+++ server/src/utils/common.py
@@ -2,15 +2,14 @@ import datetime
 
 
 def stringify_duration(
-    start_time: datetime.datetime, end_time: datetime.datetime
+    total_duration: datetime.timedelta,
 ) -> str:
     """
     Format the duration as HH:MM:SS
-    :param start_time: start time
-    :param end_time: end time
+    :param total_duration: duration as timedelta
     :return: duration as HH:MM:SS
     """
-    duration = end_time - start_time
+    duration = total_duration
     hours, remainder = divmod(duration.seconds, 3600)
     minutes, seconds = divmod(remainder, 60)
     duration_str = f"{hours:02}:{minutes:02}:{seconds:02}"